Reorganized code in main module
